users update

+++ database/migrations/2025_06_22_124757_create_item_pesanans_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
                 ->constrained()
                 ->onDelete('cascade');
             $table->integer('jumlah');
-            $table->decimal('subtotal');
+            $table->integer('subtotal');
             $table->timestamps();
         });
     }
